requirements.txt
python=3.9
pytorch=2.1.0
torchvision=0.16.0
torchaudio=2.1.0
pytorch-cuda=11.8
pytorch-lightning=2.0.7
monai=1.3.0
matplotlib=3.7.2
scipy=1.11.1
pandas=2.1.1
scikit-image=0.19.3
jsonargparse[signatures]=4.25.0
nibabel=5.1.0
wandb=0.15.12